% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merkle.R
\name{merkle_tree}
\alias{merkle_tree}
\title{Construct a Merkle tree}
\usage{
merkle_tree(hash = "sha256")
}
\arguments{
\item{hash}{The hash algorithm to use.  Default is sha-256.  Hash
functions come from the \code{openssl} package.  Valid values
are \code{sha1}, \code{sha224}, \code{sha256}, \code{sha384},
\code{sha512}, \code{blake2b}, \code{blake2s} and
\code{ripemd160}.  You \emph{could} use \code{md4}, \code{md5},
but as these are really not secure it is not recommended.}
}
\description{
Construct a Merkle tree.  At present this returns an R6 object but
this is an implementation detail subject to change.
}
